moonbeam_runtime_common

Module deal_with_fees

Structs

BlockAuthorAccountId

DealWithEthereumBaseFees

Deal with ethereum based fees. To handle tips/priority fees, use DealWithEthereumPriorityFees.

DealWithEthereumPriorityFees

Deal with ethereum based priority fees/tips. See DealWithEthereumBaseFees for base fees.

DealWithSubstrateFeesAndTip

Deal with substrate based fees and tip. This should be used with pallet_transaction_payment.

ProofSizeToFee

Type alias for converting proof size weight to fee using a constant multiplier.

WeightToFee

Combines reference time and proof size fees, charging by the more scarce resource.

Type Aliases

RefTimeToFee

Type alias for converting reference time weight to fee using a constant multiplier.