moonbeam_runtime_common

Module migrations

Common Moonbeam Migrations

Structs

DestinationAssetFeePerSecondStorageName

RemovePallet

Remove all of a pallet’s state and re-initializes it to the current in-code storage version.

ResetStorage

Reset the pallet’s storage.

Type Aliases

MultiBlockMigrations

List of common multiblock migrations to be executed by the pallet-migrations pallet. The migrations listed here are common to every moonbeam runtime.

PermanentSingleBlockMigrations

Migrations/checks that do not need to be versioned and can run on every update.

SingleBlockMigrations

All migrations that will run on the next runtime upgrade.

UnreleasedSingleBlockMigrations

Unreleased migrations. Add new ones here: