Module moonbeam_runtime_common::migrations

Migrations

Structs

AssetManagerChangeStateminePrefixes

AssetManagerPopulateAssetTypeIdStorage

AssetManagerUnitsWithAssetType

AuthorMappingAddAccountIdToNimbusLookup

This module acts as a registry where each migration is defined. Each migration should implement the “Migration” trait declared in the pallet-migrations crate. A moonbeam migration wrapping the similarly named migration in pallet-author-mapping

AuthorMappingAddKeysToRegistrationInfo

A moonbeam migration wrapping the similarly named migration in pallet-author-mapping

AuthorSlotFilterEligibleRatioToEligiblityCount

BaseFeeElasticity

BaseFee pallet, set Elasticity to zero. This migration needs to be applied before or at the same time we introduce: https://github.com/paritytech/frontier/pull/794

BaseFeePerGas

BaseFee pallet, set missing storage values.

CommonMigrations

MigrateBaseFeeElasticity

MigrateBaseFeePerGas

MigrateCollectivePallets

ParachainStakingMigrateAtStakeAutoCompound

Migrate AtStake storage item to contain 0% auto-compound values

ParachainStakingPatchIncorrectDelegationSums

Patch delegations total mismatch

ParachainStakingPurgeStaleStorage

A moonbeam migration wrapping the similarly named migration in parachain-staking

ParachainStakingSplitDelegatorStateIntoDelegationScheduledRequests

Staking split delegator state into [pallet_parachain_staking::DelegatorScheduledRequests]

SchedulerMigrationV3

XcmPaymentSupportedAssets

XcmTransactorTransactSignedWeightAndFeePerSecond